Add the Mercurial (Hg) repository changeset ident into the PDF documentation at build time.
We do this because the pdf which is built has no way to track it back and see which revision
it came from.  This does depend on the user having the repo as a checkout though,
and not a packaged copy of the source only.  This could be a problem.