Start switching over to the alternative mongo-client.
The official MongoDB C driver appears to be poorly documented, with out of date
examples, at the time this was written, and appears not to be used by many people.

Debian only appear to have packaged the alternative client.  We might look into
the official C client again in the future.